Move latrd_data context manager to tristan.data
